added SwitchAndRamp class for smoothing discontinuities
and used it in a few places in sampleplayer and looper. need to go through over time and use this class to eliminate a bunch of pops.

+++ Source/Sample.h
@@ -61,7 +61,7 @@ public:
    int LengthInSamples() const { return mNumSamples; }
    int NumChannels() const { return mData.NumActiveChannels(); }
    ChannelBuffer* Data() { return &mData; }
-   int GetPlayPosition() const { return mOffset; }
+   double GetPlayPosition() const { return mOffset; }
    void SetPlayPosition(double sample) { mOffset = sample; }
    float GetSampleRateRatio() const { return mSampleRateRatio; }
    void Reset() { mOffset = mNumSamples; }
